Porting regnet
